complete merge from trunk
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/branches/account_db@20161 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/user/FavoriteFactory.java
@@ -320,7 +320,7 @@ public class FavoriteFactory {
   }
   
   public Favorite getFavorite(User user, RecordClass recordClass, Map<String, Object> recordId) throws WdkModelException {
-    int userId = user.getUserId();
+    long userId = user.getUserId();
     String projectId = wdkModel.getProjectId();
     String rcName = recordClass.getFullName();
     String[] pkColumns = recordClass.getPrimaryKeyDefinition().getColumnRefs();
